Dynamic Instructions/pyproject.toml
[project]
name = "dynamic-instructions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.9",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
